+++ src/singleTable/model/from/fromCollection/definitions/get.ts
@@ -4,11 +4,11 @@ import type {
   SingleTableQueryParams,
 } from 'singleTable/adaptor/definitions';
 
-import type { ExtendableCollection } from 'singleTable/model';
+import type { AnyCollection } from 'singleTable/model';
 
 type BaseParams = Partial<Pick<SingleTableQueryParams<any>, 'retrieveOrder'>>;
 
-type NarrowParams<Registered extends ExtendableCollection> =
+type NarrowParams<Registered extends AnyCollection> =
   Registered['narrowBy'] extends 'RANGE_KEY'
     ? Parameters<Registered['originEntity']['getRangeKey']>[0]
     : Registered['narrowBy'] extends (...params: any[]) => DefinedNameRangeKeyConfig
@@ -24,14 +24,14 @@ type JoinParams<ParamsA, ParamsB> = ParamsA extends undefined
   : ParamsA & ParamsB;
 
 export type GetCollectionParams<
-  Registered extends ExtendableCollection,
+  Registered extends AnyCollection,
   GetParams = JoinParams<
     Parameters<Registered['getPartitionKey']>[0],
     NarrowParams<Registered>
   >,
 > = GetParams extends undefined ? [BaseParams?] : [BaseParams & GetParams];
 
-export type GetCollectionResult<Registered extends ExtendableCollection> =
+export type GetCollectionResult<Registered extends AnyCollection> =
   Registered['__type'] extends Array<any>
     ? Registered['__type']
     : Registered['__type'] | undefined;