+++ src/singleTable/model/from/fromCollection/implementation.ts
@@ -6,7 +6,7 @@ import { cascadeEval } from 'utils/conditions';
 import { SingleTableMethods, SingleTableParams } from 'singleTable/adaptor';
 import type {
   BaseJoinConfig,
-  ExtendableCollection,
+  AnyCollection,
   Extractor,
   JoinResolutionParams,
   Sorter,
@@ -138,7 +138,7 @@ export class SingleTableFromCollection<SingleParams extends SingleTableParams> {
   }: {
     item: AnyObject;
     options: AnyObject[]; // all the options available to be joined
-    join: ExtendableCollection['join'];
+    join: AnyCollection['join'];
     mapping: Record<string, AnyObject[]>; // byType to easily reduce times if joinBy is BY TYPE
   }): AnyObject {
     const joinProps = Object.fromEntries(
@@ -152,7 +152,7 @@ export class SingleTableFromCollection<SingleParams extends SingleTableParams> {
             sorter,
             extractor,
           } = config as typeof config & {
-            join?: ExtendableCollection['join'];
+            join?: AnyCollection['join'];
             sorter?: Sorter;
             extractor?: Extractor;
           };
@@ -249,7 +249,7 @@ export class SingleTableFromCollection<SingleParams extends SingleTableParams> {
 
   private cleanCollectionProps<Result extends AnyObject | AnyObject[]>(
     result: Result,
-    collection: ExtendableCollection,
+    collection: AnyCollection,
   ): Result {
     return this.cleanSingleCollectionEntry(
       result,
@@ -258,7 +258,7 @@ export class SingleTableFromCollection<SingleParams extends SingleTableParams> {
     );
   }
 
-  private buildCollection<Collection extends ExtendableCollection>(
+  private buildCollection<Collection extends AnyCollection>(
     collection: Collection,
     items: AnyObject[],
   ): GetCollectionResult<Collection> {
@@ -305,7 +305,7 @@ export class SingleTableFromCollection<SingleParams extends SingleTableParams> {
     return this.cleanCollectionProps(joined, collection);
   }
 
-  private async getPartitionCollection<Collection extends ExtendableCollection>(
+  private async getPartitionCollection<Collection extends AnyCollection>(
     collection: Collection,
     ...params: Parameters<FromCollection<Collection>['get']>
   ): Promise<GetCollectionResult<Collection>> {
@@ -344,7 +344,7 @@ export class SingleTableFromCollection<SingleParams extends SingleTableParams> {
     return this.buildCollection(collection, items);
   }
 
-  fromCollection<Collection extends ExtendableCollection>(
+  fromCollection<Collection extends AnyCollection>(
     collection: Collection,
   ): FromCollection<Collection> {
     const methods: FromCollection<Collection> = {