+++ src/singleTable/model/from/fromEntity/definitions/crud.ts
@@ -1,4 +1,4 @@
-import type { KeyParams, ExtendableSingleTableEntity } from 'singleTable/model';
+import type { KeyParams, AnyEntity } from 'singleTable/model';
 import type {
   BatchListItemsArgs,
   GetItemParams,
@@ -7,13 +7,13 @@ import type {
 } from 'provider/utils';
 import { FirstParameter, PrettifyObject, SafeObjMerge } from 'types';
 
-type NoKeyParam<Registered extends ExtendableSingleTableEntity> = FirstParameter<
+type NoKeyParam<Registered extends AnyEntity> = FirstParameter<
   Registered['getKey']
 > extends undefined
   ? true
   : false;
 
-export type EntityGetParams<Registered extends ExtendableSingleTableEntity> =
+export type EntityGetParams<Registered extends AnyEntity> =
   NoKeyParam<Registered> extends true
     ? [Omit<GetItemParams<Registered['__entity']>, 'table' | 'key'>?]
     : [
@@ -21,14 +21,14 @@ export type EntityGetParams<Registered extends ExtendableSingleTableEntity> =
           KeyParams<Registered>,
       ];
 
-export type EntityBatchGetParams<Registered extends ExtendableSingleTableEntity> = Omit<
+export type EntityBatchGetParams<Registered extends AnyEntity> = Omit<
   BatchListItemsArgs<Registered['__entity']>,
   'table' | 'keys'
 > & {
   keys: Array<KeyParams<Registered>>;
 };
 
-export type DeleteEntityParams<Registered extends ExtendableSingleTableEntity> =
+export type DeleteEntityParams<Registered extends AnyEntity> =
   NoKeyParam<Registered> extends true
     ? [Omit<DeleteItemParams<Registered['__entity']>, 'table' | 'key'>?]
     : [
@@ -36,17 +36,17 @@ export type DeleteEntityParams<Registered extends ExtendableSingleTableEntity> =
           Omit<DeleteItemParams<Registered['__entity']>, 'table' | 'key'>,
       ];
 
-export type CreateEntityParams<Registered extends ExtendableSingleTableEntity> =
+export type CreateEntityParams<Registered extends AnyEntity> =
   Parameters<Registered['getCreationParams']>;
 
-export type UpdateEntityParams<Registered extends ExtendableSingleTableEntity> = Omit<
+export type UpdateEntityParams<Registered extends AnyEntity> = Omit<
   UpdateParams<Registered['__entity']>,
   'table' | 'key'
 > &
   KeyParams<Registered>;
 
 export type UpdateReturn<
-  Entity extends ExtendableSingleTableEntity,
+  Entity extends AnyEntity,
   Params extends UpdateEntityParams<Entity>,
 > = Params['returnUpdatedProperties'] extends true
   ? PrettifyObject<