+++ src/singleTable/model/from/fromEntity/definitions/query/partition.ts
@@ -1,24 +1,24 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import type { ExtendableSingleTableEntity } from 'singleTable/model';
+import type { AnyEntity } from 'singleTable/model';
 import type { SingleTableQueryParams } from 'singleTable/adaptor/definitions';
 import type { QueryResult } from 'provider';
 import type { FirstParameter, HasDefined, OptionalTupleIf, SafeObjMerge } from 'types';
 import type { OptionalTupleIfUndefined, QueryConfigParams } from './common';
 
-type BaseQueryParams<Registered extends ExtendableSingleTableEntity> = SafeObjMerge<
+type BaseQueryParams<Registered extends AnyEntity> = SafeObjMerge<
   QueryConfigParams<Registered['__entity']>,
   FirstParameter<Registered['getPartitionKey']>
 >;
 
-type CustomQueryParams<Registered extends ExtendableSingleTableEntity> =
+type CustomQueryParams<Registered extends AnyEntity> =
   BaseQueryParams<Registered> &
     Pick<SingleTableQueryParams<Registered['__entity']>, 'range'>;
 
-type EntityQueryResult<Registered extends ExtendableSingleTableEntity> = QueryResult<
+type EntityQueryResult<Registered extends AnyEntity> = QueryResult<
   Registered['__entity']
 >;
 
-type RangeQueries<Registered extends ExtendableSingleTableEntity> = Registered extends {
+type RangeQueries<Registered extends AnyEntity> = Registered extends {
   rangeQueries: any;
 }
   ? {
@@ -42,7 +42,7 @@ type RangeQueries<Registered extends ExtendableSingleTableEntity> = Registered e
     }
   : unknown;
 
-export type PartitionQueryMethods<Registered extends ExtendableSingleTableEntity> = {
+export type PartitionQueryMethods<Registered extends AnyEntity> = {
   custom(
     ...params: OptionalTupleIfUndefined<
       FirstParameter<Registered['getPartitionKey']>,