+++ changelog.md
@@ -4,6 +4,8 @@
 
 - [BREAKING] *UUID*: Remove `uuid` library in favor of crypto's `randomUUID`. Min node supported version is now v16.
 - [BREAKING] `schema.fromCollection` and `schema.fromEntity` generators removed. Use `schema.from(xxx)` for both!
+- [BREAKING - Types] `ExtendableCollection` type renamed to `AnyCollection`
+- [Breaking - Types] `ExtendableSingleTableEntity` type renamed to `AnyEntity`
 - *Feature*: Direct `collection` creation from `partition.collection()`
 - *Feature*: `schema.from(xxx).update()` now property infers return type if `returnUpdatedProperties` is true.
 - *Feature*: Entities' **Range Queries** now support definitions with only the operation provided. Required params will be the default for its respective operation